requirements.txt
Django
django-appconf==1.0.2
django-imagekit==4.0.2
djangorestframework==3.8.2
djangorestframework-gis==0.13
ExifRead==2.1.2
#pygdal
#GDAL
numpy==1.15.1
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
six==1.11.0
drfapikey
rest_condition
piexif
django-dirtyfields
django-filter
